move CleverLibrary::Api::Client to CleverIntegration::LibraryClient
